1.0.14 1.删除DeskPosLoginActivity中不用的代码 2.将DeskPosMainActivity中的累计消费金额显示修改为double类型 3.DishAdapter里菜品单价显示，采用double类型 4.给HttpUtil通信增加日志 5.通信的float类型，全部转为double类型

+++ app/src/main/java/com/rankway/controller/activity/project/DeskPosPayMainActivity.java
@@ -613,7 +613,7 @@ public class DeskPosPayMainActivity
 
         tvTotalCount.setText("总数：" + shiftEntity.getTotalCount());
 
-        float fTotalAmount = (float) (shiftEntity.getTotalAmount() * 0.01);
+        double fTotalAmount = (double) (shiftEntity.getTotalAmount() * 0.01);
         tvTotalAmount.setText(String.format("总金额：%.2f", fTotalAmount));
     }
